Rename app -> my_app
* Zappa needs app entry function and can't find app.app
